+++ frontend/components/cart/individual_cart_item.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useDispatch } from "react-redux";
 import { Link } from "react-router-dom";
-import { updateProductInCart } from "../../actions/cart_actions";
+import { updateProductInCart } from "../../actions/cartActions";
 
 const IndividualCart = ({
   cartProduct,
